% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tides.R
\name{tides_predictions}
\alias{tides_predictions}
\title{Fetch tide predictions}
\usage{
tides_predictions(
  station,
  begin_date,
  end_date,
  datum = "MLLW",
  interval = "hilo",
  units = "metric"
)
}
\arguments{
\item{station}{Station ID}

\item{begin_date}{Start date (Date or "YYYY-MM-DD")}

\item{end_date}{End date (Date or "YYYY-MM-DD")}

\item{datum}{Vertical datum: "MLLW" (default), "MSL"}

\item{interval}{"hilo" for high/low only, "6" for 6-minute, "h" for hourly}

\item{units}{"metric" (default) or "english"}
}
\value{
tibble: datetime, value
}
\description{
Retrieves tide predictions (high/low or 6-minute intervals) for a station.
}
