% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tides.R
\name{tides_stations}
\alias{tides_stations}
\title{Fetch list of NOAA tide stations}
\usage{
tides_stations()
}
\value{
tibble: id, name, state, lat, lng
}
\description{
Returns metadata for all active water level stations.
}
